Adiconando bootstrap, gerenciamentos de imoveis, e banco de dados

+++ back-end/app/Http/Controllers/ImmobileController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 use App\Models\Immobile;
 use App\Models\VisitRequest;
 use App\Models\User;
+use App\Mail\VisitRequestMail;
+use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Auth;
 
 class ImmobileController extends Controller
@@ -50,26 +52,29 @@ class ImmobileController extends Controller
 
     public function destroy($id)
     {
-        $immobile = Immobile::findOrFail($id);
+        $immobile = Immobile::find($id);
+
+        if (!$immobile) {
+            return response()->json(['message' => 'Imóvel não encontrado'], 404);
+        }
+
         $immobile->delete();
 
         return response()->json(['message' => 'Imóvel deletado com sucesso'], 200);
     }
 
+
     public function show($id)
     {
         $immobile = Immobile::findOrFail($id);
 
+        $user = Auth::user();
+        if ($immobile->user_id !== $user->id) {
+            return response()->json(['message' => 'Acesso não autorizado'], 403);
+        }
+
         return response()->json($immobile, 200);
     }
-
-    public function index()
-    {
-        $user = Auth::user();
-        $imoveis = $user->immobile;
-        return response()->json(['imoveis' => $imoveis], 200);
-    }
-
 
     public function markVisit(Request $request, $id)
     {
@@ -81,17 +86,21 @@ class ImmobileController extends Controller
         ]);
 
         $immobile = Immobile::findOrFail($id);
+        $user_id = 1; // Defina o ID do usuário fixo aqui
 
         $visitRequest = new VisitRequest($validatedData);
         $visitRequest->immobile()->associate($immobile);
+        $visitRequest->user_id = $user_id; // Defina o ID do usuário fixo aqui
         $visitRequest->save();
 
         return response()->json($visitRequest, 201);
     }
+    
 
     public function getAllVisitRequests()
     {
-        $visitRequests = VisitRequest::with('immobile')->get();
+        $visitRequests = VisitRequest::with('immobile')
+            ->get();
 
         $formattedVisitRequests = $visitRequests->map(function ($visitRequest) {
             return [
@@ -106,5 +115,11 @@ class ImmobileController extends Controller
         });
 
         return response()->json($formattedVisitRequests);
+    }
+    public function index()
+    {
+        $user = Auth::user();
+        $imoveis = $user->immobile;
+        return response()->json(['imoveis' => $imoveis], 200);
     }
 }
